+++ tiger-testenv-mgr/src/frontend/src/components/serverlog/ServerLog.vue
@@ -18,7 +18,7 @@
           <td>
             <div class="ps-0 pt-2 pb-4">
             <div class="justify-content-between" id="test-server-log-pane-buttons" style="display: inline-block;" >
-              <div class="btn active server-buttons test-server-log-pane-server-all" @click="setServer(selectedServers, '__all__', $event)">Show all logs</div>
+              <div class="btn active server-buttons" id="test-server-log-pane-server-all" @click="setServer(selectedServers, '__all__', $event)">Show all logs</div>
               <div v-for="(serverName,logIndex) in logServers" :key="logIndex" style="display: inline-block;">
                 <div class="btn server-buttons" :id="`test-server-log-pane-server-${serverName}`" @click="setServer(selectedServers, serverName, $event)">{{ serverName }}</div>
               </div>