+++ tiger-uitests/src/test/java/de/gematik/test/tiger/playwright/workflowui/TPauseTests.java
@@ -3,6 +3,7 @@
  */
 
 package de.gematik.test.tiger.playwright.workflowui;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import org.junit.jupiter.api.AfterEach;
@@ -18,10 +19,10 @@ class TPauseTests extends AbstractTests {
         page.querySelector("#test-sidebar-pause-icon").click();
         assertAll(
             () -> assertThat(page.locator("#sidebar-left.test-sidebar-paused").isVisible()).isTrue()
-            // cannot check pending because tests when through or have not startet jet
-            //() -> assertThat(page.locator("#execution_table .test-status-pending").isVisible()).isTrue(),
-            //() -> assertThat(page.locator("#execution_table .test-pending").isVisible()).isTrue()
+          //  () -> assertThat(page.locator("#execution_table .test-pending").first().isVisible()).isTrue()
         );
+
+        screenshot(page, "sidebar_pause.png");
     }
 
     @AfterEach