+++ tiger-uitests/src/test/java/de/gematik/test/tiger/playwright/workflowui/XDynamicMainContentTests.java
@@ -6,7 +6,12 @@ package de.gematik.test.tiger.playwright.workflowui;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInfo;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 /**
  * Tests for dynamic content of main content area, e.g. server log pane.
@@ -23,6 +28,43 @@ class XDynamicMainContentTests extends AbstractTests {
         );
     }
 
+    @ParameterizedTest
+    @ValueSource(strings = {"localTigerProxy", "remoteTigerProxy", "httpbin"})
+    void testServerLogsLogsOfServerShown(String server) {
+        page.querySelector("#test-server-log-tab").click();
+        page.querySelector("#test-server-log-pane-server-" + server).click();
+        page.locator(".test-server-log-pane-log-1").all()
+            .forEach(log -> assertThat(log.textContent().equals(server)));
+    }
+
+    @Test
+    void testServerLogNoLogsOfHttpbinShown() {
+        page.querySelector("#test-server-log-tab").click();
+        page.querySelector("#test-server-log-pane-server-localTigerProxy").click();
+        page.querySelector("#test-server-log-pane-server-remoteTigerProxy").click();
+        page.locator(".test-server-log-pane-log-1").all()
+            .forEach(log -> assertThat(log.textContent()).isNotEqualTo("httpbin"));
+    }
+
+    @Test
+    void testServerLogNoLogsShown() {
+        page.querySelector("#test-server-log-tab").click();
+        page.querySelector("#test-server-log-pane-server-all").click();
+        page.locator("#test-server-log-pane-input-text").type("ready");
+        assertThat(page.locator(".test-server-log-pane-log-1").isVisible()).isFalse();
+        page.locator("#test-server-log-pane-input-text").type("");
+    }
+
+//    @Test
+//    void testServerLogTwoLogsShown() {
+//        page.querySelector("#test-server-log-tab").click();
+//        page.querySelector("#test-server-log-pane-server-all").click();
+//        page.locator("#test-server-log-pane-input-text").type("READY");
+//        assertThat(page.locator(".test-server-log-pane-log-1").all().size()).isEqualTo(2);
+//        page.locator("#test-server-log-pane-input-text").type("");
+//    }
+
+
     @Test
     void testClickOnRequestOpensRbelLogDetails() {
         page.querySelector("#test-execution-pane-tab").click();
@@ -31,4 +73,12 @@ class XDynamicMainContentTests extends AbstractTests {
         assertThat(page.locator("#rbellog_details_pane").isVisible()).isTrue();
     }
 
+    @BeforeEach
+    void printInfoStarted(TestInfo testInfo) {
+        System.out.println("started = " + testInfo.getDisplayName());
+    }
+    @AfterEach
+    void printInfoFinished(TestInfo testInfo) {
+        System.out.println("finished = " + testInfo.getDisplayName());
+    }
 }