+++ tiger-uitests/src/test/java/de/gematik/test/tiger/playwright/workflowui/ZQuitTests.java
@@ -4,38 +4,176 @@
 
 package de.gematik.test.tiger.playwright.workflowui;
 
-import static java.lang.Thread.sleep;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
 import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import com.microsoft.playwright.Page;
+import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 
 /**
  * These tests should run at the very last because the testQuitButton() quits the tiger/workflowui.
  */
 @Slf4j
+@TestMethodOrder(OrderAnnotation.class)
 class ZQuitTests extends AbstractTests {
 
+    @BeforeEach
+    void printInfoStarted(TestInfo testInfo) {
+        System.out.println("started = " + testInfo.getDisplayName());
+    }
+    @AfterEach
+    void printInfoFinished(TestInfo testInfo) {
+        System.out.println("finished = " + testInfo.getDisplayName());
+    }
+
     @Test
-    void testResetButton() throws InterruptedException {
+    @Order(1)
+    void testClickOnLastRequestChangesPageNumberInRbelLogDetails() throws InterruptedException {
+        page.querySelector("#test-execution-pane-tab").click();
+        page.locator("#test-webui-slider").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#dropdown-page-selection").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#pageSelector .dropdown-item").first().click();
+        page.frameLocator("#rbellog-details-iframe").locator("#dropdown-page-size").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#sizeSelector .dropdown-item").nth(1).click();
+        page.locator("#test-webui-slider").click();
+        assertThat(page.locator("#rbellog_details_pane").isVisible()).isFalse();
+        page.locator(".test-rbel-link").first().click();
+        String number1 = page.locator(".test-rbel-link").first().textContent();
+        String number2 = page.locator(".test-rbel-link").last().textContent();
+        assertAll(
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator(".test-message-number").first().textContent()).isEqualTo(number1),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator(".test-message-number").last().textContent()).isNotEqualTo(number2)
+        );
+        String pageNo = page.frameLocator("#rbellog-details-iframe").locator("#pageNumberDisplay").textContent();
+        page.locator("#test-webui-slider").click();
+        assertThat(page.locator("#rbellog_details_pane").isVisible()).isFalse();
+        page.locator(".test-rbel-link").last().click();
+        // somehow I need to wait
+        await()
+            .atMost(10, TimeUnit.SECONDS)
+            .untilAsserted(() -> assertNotNull(page.frameLocator("#rbellog-details-iframe").locator("#pageNumberDisplay")));
+        String pageNo2 = page.frameLocator("#rbellog-details-iframe").locator("#pageNumberDisplay").textContent();
+        int value = Integer.parseInt(number2)+1;
+        assertAll(
+            () -> assertThat(pageNo).isNotEqualTo(pageNo2),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator(".test-message-number").first().textContent()).isNotEqualTo(number1),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator(".test-message-number").last().textContent()).isEqualTo(String.valueOf(value))
+        );
+    }
+
+    @Test
+    @Order(2)
+    void testFilterModal() {
+        page.querySelector("#test-execution-pane-tab").click();
+        page.locator("#test-webui-slider").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#filterModalBtn").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#resetFilterCriterionBtn").click();
+        assertAll(
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#filterModalDialog").isVisible()).isTrue(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#setFilterCriterionInput").isVisible()).isTrue(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#setFilterCriterionInput").textContent()).isEmpty(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#requestFromContent").isVisible()).isTrue(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#requestToContent").isVisible()).isTrue(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#requestToContent").textContent()).contains("no request"),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#requestFromContent").textContent()).contains("no request"),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#resetFilterCriterionBtn").isVisible()).isTrue(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#setFilterCriterionBtn").isVisible()).isTrue(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#filteredMessage").isVisible()).isTrue(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#filterModalButtonClose").isVisible()).isTrue(),
+            () -> assertThat(page.frameLocator("#rbellog-details-iframe").locator("#filteredMessage").textContent()).isEqualTo("Filter didn\'t match any of the 64 messages.")
+        );
+        page.frameLocator("#rbellog-details-iframe").locator("#filterModalButtonClose").click();
+        assertThat(page.frameLocator("#rbellog-details-iframe").locator("#filterModalDialog").isVisible()).isFalse();
+    }
+
+    @Test
+    @Order(3)
+    void testFilterModalResetFilter() {
+        page.querySelector("#test-execution-pane-tab").click();
+        page.locator("#test-webui-slider").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#filterModalBtn").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#setFilterCriterionInput").fill("$.body == \"hello=world\"");
+        page.frameLocator("#rbellog-details-iframe").locator("#setFilterCriterionBtn").click();
+        await()
+            .atMost(10, TimeUnit.SECONDS)
+            .untilAsserted(() -> assertNotNull(page.frameLocator("#rbellog-details-iframe").locator("#filteredMessage")));
+        String filteredMessage = page.frameLocator("#rbellog-details-iframe").locator("#filteredMessage").textContent();
+        int count = page.frameLocator("#rbellog-details-iframe").locator("#test-rbel-section .test-msg-body-content").count();
+        page.frameLocator("#rbellog-details-iframe").locator("#resetFilterCriterionBtn").click();
+        await()
+            .atMost(10, TimeUnit.SECONDS)
+            .untilAsserted(() -> assertNotNull( page.frameLocator("#rbellog-details-iframe").locator("#filterModalButtonClose")));
+        page.frameLocator("#rbellog-details-iframe").locator("#filterModalButtonClose").click();
+        assertThat(filteredMessage).isEqualTo("4 of 64 did match the filter criteria.");
+        assertThat(count).isEqualTo(3);
+        assertThat(page.frameLocator("#rbellog-details-iframe").locator("#filteredMessage").textContent()).isEqualTo("Filter didn't match any of the 64 messages.");
+    }
+
+    @Test
+    @Order(4)
+    void testFilterModalSetSenderFilter() {
+        page.querySelector("#test-execution-pane-tab").click();
+        page.locator("#test-webui-slider").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#filterModalBtn").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#setFilterCriterionInput").fill("$.sender == \"put\"");
+        page.frameLocator("#rbellog-details-iframe").locator("#setFilterCriterionBtn").click();
+        await()
+            .atMost(10, TimeUnit.SECONDS)
+            .untilAsserted(() -> assertNotNull(page.frameLocator("#rbellog-details-iframe").locator("#requestToContent")));
+        String requestToContent = page.frameLocator("#rbellog-details-iframe").locator("#requestToContent").textContent();
+        String requestFromContent = page.frameLocator("#rbellog-details-iframe").locator("#requestFromContent").textContent();
+        String filteredMessage = page.frameLocator("#rbellog-details-iframe").locator("#filteredMessage").textContent();
+        page.frameLocator("#rbellog-details-iframe").locator("#resetFilterCriterionBtn").click();
+        page.frameLocator("#rbellog-details-iframe").locator("#filterModalButtonClose").click();
+        assertAll(
+            () -> assertThat(requestToContent).contains("no request"),
+            () -> assertThat(requestFromContent).contains("no request"),
+            () -> assertThat(filteredMessage).isEqualTo("0 of 64 did match the filter criteria.")
+        );
+    }
+
+    @Test
+    @Order(5)
+    void testResetButton() {
         page.querySelector("#test-execution-pane-tab").click();
         page.locator("#test-webui-slider").click();
 
         Page externalPage = page.waitForPopup(() -> page.locator("#test-rbel-webui-url").click());
-        sleep(1000);
+        await()
+            .atMost(10, TimeUnit.SECONDS)
+            .untilAsserted(() -> assertNotNull(externalPage.locator("#rbelmsglist .test-card")));
         assertThat(externalPage.locator("#rbelmsglist .test-card").count()).isPositive();
         externalPage.locator("#resetMsgs").click();
         assertThat(externalPage.locator("#rbelmsglist .test-card").count()).isZero();
+        externalPage.close();
     }
 
     @Test
+    @Order(7)
     void testQuitButton() {
         page.querySelector("#test-tiger-logo").click();
         page.querySelector("#test-sidebar-quit-icon").click();
         assertAll(
             () -> assertThat(page.querySelector("#sidebar-left.test-sidebar-quit").isVisible()).isTrue(),
             () -> assertThat(page.querySelector("#workflow-messages.test-messages-quit").isVisible()).isTrue()
+        );
+
+        page.screenshot(new Page.ScreenshotOptions().setFullPage(false).setPath(getPath("workflowui_quit.png")));
+    }
+
+    @Test
+    @Order(6)
+    void testFeatures() {
+        page.querySelector("#test-tiger-logo").click();
+        assertAll(
+            () -> assertThat(page.locator("#test-sidebar-featurelistbox").isVisible()).isTrue(),
+            () -> assertThat(page.locator("#test-sidebar-featurelistbox .test-sidebar-feature-name").count()).isEqualTo(1),
+            () -> assertThat(page.locator("#test-sidebar-featurelistbox .test-sidebar-scenario-name").count()).isEqualTo(25),
+            () -> assertThat(page.locator("#test-sidebar-featurelistbox .test-sidebar-scenario-index").count()).isEqualTo(11)
         );
     }
 }